+++ web-console/src/app/(root)/(authenticated)/layout.tsx
@@ -1,15 +1,14 @@
 'use client'
 
-import { useAuthentication } from '$lib/compositions/useAuth'
+import { useAuth } from '$lib/compositions/auth/useAuth'
 import { redirect } from 'next/navigation'
-
-import type { ReactNode } from 'react'
+import { ReactNode } from 'react'
 
 export default (props: { children: ReactNode }) => {
-  const auth = useAuthentication()
-
-  if (!auth && (process.env.NODE_ENV !== 'development' || process.env.NEXT_PUBLIC_WEB_CONSOLE_TEST_AUTH)) {
+  const { auth } = useAuth()
+  if (auth === 'Unauthenticated') {
     redirect('/login')
   }
+
   return props.children
 }