+++ web-console/src/lib/services/manager/models/AuthProvider.ts
@@ -1,0 +1,15 @@
+/* generated using openapi-typescript-codegen -- do no edit */
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { ProviderAwsCognito } from './ProviderAwsCognito'
+import type { ProviderGoogleIdentity } from './ProviderGoogleIdentity'
+
+export type AuthProvider =
+  | {
+      AwsCognito: ProviderAwsCognito
+    }
+  | {
+      GoogleIdentity: ProviderGoogleIdentity
+    }